services/workflow-orchestrator-service/requirements.txt
fastapi[all]==0.115.8
pydantic-settings==2.6.1
uvicorn[standard]==0.32.1
